[component/video] add mute function
