+++ src/components/Video.js
@@ -1,4 +1,4 @@
-import React, { useState, useRef } from "react"
+import React, { useState} from "react"
 import VideoFooter from "./VideoFooter"
 import "../styles/Video.css"
 import SideBar from "./SideBar"
@@ -12,59 +12,47 @@ import Loading from "./Loading/Loading"
 import Error from "./Error"
 
 const Video = (props) => {
-  const queryVideos = gql`
-    {
-      videos {
-	      id
-        description
-        userId {
-          id
-          username
-        }
-        videoUrl
-        likes
-        shares
+  let [muted, setMuted] = useState(true)
+  let video = props.video
+
+    const handleMute = () => {
+      if(muted) {
+        setMuted(false)
+      } else {
+        setMuted(true)
       }
     }
-  `
 
-  return (
-    <div className="videos">
-      <TopNav />
-      <Query query={queryVideos}>
-        {({ loading, error, data }) => {
-	        if (loading) return <Loading />;
-	        if (error) return <Error errorMessage={error.message} />;
+    const handleUnmute = () => {
+      console.log('scrolling')
+    }
 
-          let videos = data.videos.map((video) => {
-            return (
-              <div className="Video-Container" key={video.id}>
-                <SideBar
-	                videoId={video.id}
-                  userId={video.userId.id}
-                  likes={video.likes}
-                  shares={video.shares}
-                />
-                <VideoFooter
-                  username={video.userId.username}
-                  description={video.description}
-                />
-                <video
-                  className="videoPlayer"
-                  src={`https://res.cloudinary.com/agregis/video/upload/c_scale,g_south_west,l_logo_transparent_wvrps0,w_144/${video.videoUrl}.mp4`}
-                  autoPlay={true}
-                  loop={true}
-                  muted
-                />
-              </div>
-            )
-          })
-
-          return videos
-        }}
-      </Query>
-    </div>
-  )
+    return (
+      <div className="videos">
+        <TopNav />
+        <div className="Video-Container" key={video.id}>
+          <SideBar
+            videoId={video.id}
+            userId={video.userId.id}
+            likes={video.likes}
+            shares={video.shares}
+          />
+          <VideoFooter
+            username={video.userId.username}
+            description={video.description}
+          />
+          <video
+            className="videoPlayer"
+            src={`https://res.cloudinary.com/agregis/video/upload/c_scale,g_south_west,l_logo_transparent_wvrps0,w_144/${video.videoUrl}.mp4`}
+            loop={true}
+            autoPlay
+            muted={muted}
+            onClick={handleMute}
+            onScroll={handleUnmute}
+          />
+        </div>
+      </div>
+    )
 }
 
 export default Video
